requirements.txt
git+https://github.com/openai/CLIP.git
fastapi
httpx
natsort
numpy
Pillow-SIMD
pydantic
pymilvus
pymongo
PyYAML
regex
skimage
tqdm
uvicorn
tensorflow
torch
paddleocr
paddlepaddle-gpu